--- src/main/java/com/linerup/lineup_backend/common/model/PostStatus.java
@@ -1,14 +1,0 @@
-package com.linerup.lineup_backend.common.model;
-
-/**
- * packageName    : com.linerup.lineup_backend.common.model
- * fileName       : PostStatus
- * author         : moongi
- * date           : 12/8/23
- * description    :
- */
-public enum PostStatus {
-    CLOSED, // 구인완료
-    ACTIVE, // 구인중
-    EXPIRED // 기간만료
-}
