--- src/main/java/com/linerup/lineup_backend/domain/repository/UserRepository.java
@@ -1,14 +1,0 @@
-package com.linerup.lineup_backend.domain.repository;
-
-
-import com.linerup.lineup_backend.member.command.domain.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<Member, Long> {
-
-  Optional<Member> findByProviderId(String providerId);
-}
