+++ src/main/java/com/linerup/lineup_backend/entity/member/Role.java
@@ -1,0 +1,13 @@
+package com.linerup.lineup_backend.entity.member;
+/**
+* @author :
+* @version : 1.0.0
+* @package : com.linerup.lineup_backend.domain
+* @name : Role.java
+* @date : 2023/08/25 2:53 AM
+* @modified :
+**/
+public enum Role {
+  USER,
+  ADMIN
+}
