+++ src/main/java/com/linerup/lineup_backend/entity/post/Category.java
@@ -1,0 +1,17 @@
+package com.linerup.lineup_backend.entity.post;
+
+/**
+ * packageName    : com.linerup.lineup_backend.common.model
+ * fileName       : Category
+ * author         : moongi
+ * date           : 12/8/23
+ * description    :
+ */
+public enum Category {
+    CUSTOM, // 맞춤
+    EXPRESS, // 급줄
+    DEPARTMENT, // 백화점
+    RESTAURANT, // 음식점
+    POPUP, // 팝업스토어
+    CAFE, // 카페
+}
