+++ src/main/java/com/linerup/lineup_backend/entity/post/PostStatus.java
@@ -1,0 +1,14 @@
+package com.linerup.lineup_backend.entity.post;
+
+/**
+ * packageName    : com.linerup.lineup_backend.common.model
+ * fileName       : PostStatus
+ * author         : moongi
+ * date           : 12/8/23
+ * description    :
+ */
+public enum PostStatus {
+    CLOSED, // 구인완료
+    ACTIVE, // 구인중
+    EXPIRED // 기간만료
+}
