+++ src/main/java/com/linerup/lineup_backend/entity/post/RecruitAge.java
@@ -1,0 +1,16 @@
+package com.linerup.lineup_backend.entity.post;
+
+/**
+ * packageName    : com.linerup.lineup_backend.entity.post
+ * fileName       : RecruitAge
+ * author         : moongi
+ * date           : 12/10/23
+ * description    :
+ */
+public enum RecruitAge {
+ TWENTIES,
+ THIRTIES,
+ FORTIES,
+ FIFTIES,
+ SIXTIES,
+}
