--- src/main/java/com/linerup/lineup_backend/member/command/application/MemberService.java
@@ -1,15 +1,0 @@
-package com.linerup.lineup_backend.member.command.application;
-
-import org.springframework.stereotype.Service;
-
-/**
- * packageName    : com.linerup.lineup_backend.domain.member
- * fileName       : MemberService
- * author         : moongi
- * date           : 11/29/23
- * description    :
- */
-
-@Service
-public class MemberService {
-}
