--- src/main/java/com/linerup/lineup_backend/member/command/domain/MemberRepository.java
@@ -1,15 +1,0 @@
-package com.linerup.lineup_backend.member.command.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-/**
- * packageName    : com.linerup.lineup_backend.domain.member
- * fileName       : MemberRepository
- * author         : moongi
- * date           : 11/29/23
- * description    :
- */
-
-public interface MemberRepository extends JpaRepository<Member, Long> {
-
-}
