--- src/main/java/com/linerup/lineup_backend/post/command/application/PostService.java
@@ -1,14 +1,0 @@
-package com.linerup.lineup_backend.post.command.application;
-
-import org.springframework.stereotype.Service;
-
-/**
- * packageName    : com.linerup.lineup_backend.domain.post
- * fileName       : PostService
- * author         : moongi
- * date           : 11/29/23
- * description    :
- */
-@Service
-public class PostService {
-}
