+++ src/main/java/com/linerup/lineup_backend/service/MemberService.java
@@ -1,0 +1,14 @@
+package com.linerup.lineup_backend.service;
+
+import org.springframework.stereotype.Service;
+
+/**
+ * packageName    : com.linerup.lineup_backend.domain.member
+ * fileName       : MemberService
+ * author         : moongi
+ * date           : 11/29/23
+ * description    :
+ */
+@Service
+public class MemberService {
+}
