+++ src/main/java/com/linerup/lineup_backend/service/PostService.java
@@ -1,0 +1,14 @@
+package com.linerup.lineup_backend.service;
+
+import org.springframework.stereotype.Service;
+
+/**
+ * packageName    : com.linerup.lineup_backend.domain.post
+ * fileName       : PostService
+ * author         : moongi
+ * date           : 11/29/23
+ * description    :
+ */
+@Service
+public class PostService {
+}
